+++ crates/rpc-ws-server/src/lib.rs
@@ -1,3 +1,0 @@
-pub mod notify_controller;
-pub mod server;
-pub mod subscription;