+++ crates/utils/src/lib.rs
@@ -2,6 +2,7 @@ pub mod exponential_backoff;
 pub mod fee;
 pub mod genesis_info;
 pub mod local_cells;
+pub mod polyjuice_parser;
 mod query_rollup_cell;
 pub mod script_log;
 pub mod since;